requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pymupdf>=1.24.14
google-generativeai>=0.8.0
pillow>=11.0.0
python-dotenv>=1.0.0
httpx>=0.28.0
